Update setup.py
Small adjustment to the spelling of the env variables.

Read the .env from the root instead of from the /website subfolder that this script is in

Allow administrator to have their newly input discord id read for the setup if they made a mistake the first time.  It will not be ignored.